add login ws functionality.

+++ src/main/java/com/d4d/controller/SchoolController.java
@@ -7,6 +7,7 @@ package com.d4d.controller;
 
 import com.d4d.model.School;
 import com.d4d.service.SchoolService;
+import com.d4d.util.ApiKey;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -28,20 +29,23 @@ import org.springframework.web.bind.annotation.RestController;
 public class SchoolController {
 
     //for Token based authentication
-    private String apiKey = "1234";
-
+    @Autowired
+    private ApiKey apiKey;
+    
     @Autowired
     private SchoolService schoolService;
 
     @ResponseBody
-    @RequestMapping(value = {"/create"}, method = RequestMethod.POST, consumes = "application/json")    
+    @RequestMapping(value = {"/create"}, method = RequestMethod.POST, consumes = "application/json")
     public ResponseEntity<String> create(@RequestHeader("apikey") String api_key, @RequestBody School school) {
 
-        if (api_key.equals(apiKey)) {
+        if (apiKey.checkApiKey(api_key)) {
 
             try {
+
                 School p = schoolService.create(school);
                 return new ResponseEntity<>("School inserted successfully with id: " + p.getId(), HttpStatus.OK);
+
             } catch (Exception e) {
                 e.printStackTrace();
                 return new ResponseEntity<>("error when adding new School", HttpStatus.EXPECTATION_FAILED);
@@ -53,32 +57,91 @@ public class SchoolController {
     }
 
     @RequestMapping("/get")
-    public School getSchool(@RequestParam String name) {
-        return schoolService.getByName(name);
+    public ResponseEntity<School> getSchool(@RequestHeader("apikey") String api_key, @RequestParam String name) {
+        if (apiKey.checkApiKey(api_key)) {
+            try {
+
+                School school = schoolService.getByName(name);
+                return new ResponseEntity<>(school, HttpStatus.OK);
+
+            } catch (Exception e) {
+                e.printStackTrace();
+                return new ResponseEntity<>(HttpStatus.EXPECTATION_FAILED);
+            }
+        } else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @RequestMapping("/getAll")
-    public List<School> getAll() {
-        return schoolService.getAll();
+    public ResponseEntity<List<School>> getAll(@RequestHeader("apikey") String api_key) {
+
+        if (apiKey.checkApiKey(api_key)) {
+            try {
+                List<School> schools = schoolService.getAll();
+                return new ResponseEntity<>(schools, HttpStatus.OK);
+
+            } catch (Exception e) {
+                e.printStackTrace();
+                return new ResponseEntity<>(HttpStatus.EXPECTATION_FAILED);
+            }
+        } else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @ResponseBody
     @RequestMapping(value = {"/update"}, method = RequestMethod.POST, consumes = "application/json")
-    public String update(@RequestHeader("apikey") String api_key, @RequestBody School school) {
-        School p = schoolService.update(school);
-        return p.toString();
+    public ResponseEntity<String> update(@RequestHeader("apikey") String api_key, @RequestBody School school) {
+
+        if (apiKey.checkApiKey(api_key)) {
+            try {
+
+                School p = schoolService.update(school);
+                return new ResponseEntity<>(p.toString(), HttpStatus.OK);
+
+            } catch (Exception e) {
+                e.printStackTrace();
+                return new ResponseEntity<>(HttpStatus.EXPECTATION_FAILED);
+            }
+        } else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @RequestMapping("/delete")
-    public String delete(@RequestParam String firstName
-    ) {
-        schoolService.delete(firstName);
-        return "Deleted " + firstName;
+    public ResponseEntity<String> delete(@RequestHeader("apikey") String api_key, @RequestParam String name) {
+
+        if (apiKey.checkApiKey(api_key)) {
+            try {
+
+                schoolService.delete(name);
+                return new ResponseEntity<>("Deleted " + name, HttpStatus.OK);
+
+            } catch (Exception e) {
+                e.printStackTrace();
+                return new ResponseEntity<>(HttpStatus.EXPECTATION_FAILED);
+            }
+        } else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @RequestMapping("/deleteAll")
-    public String deleteAll() {
-        schoolService.deleteAll();
-        return "All Deleted";
+    public ResponseEntity<String> deleteAll(@RequestHeader("apikey") String api_key) {
+
+        if (apiKey.checkApiKey(api_key)) {
+            try {
+
+                schoolService.deleteAll();
+                return new ResponseEntity<>("All Schools Deleted", HttpStatus.OK);
+
+            } catch (Exception e) {
+                e.printStackTrace();
+                return new ResponseEntity<>(HttpStatus.EXPECTATION_FAILED);
+            }
+        } else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
 }
